requirements.txt
streamlit
cohere
groq
pypdf
langchain
langchain-core
python-dotenv
pinecone-client